Switch to Pytest as default runner
