+++ elasticsearch/helpers/test.py
@@ -48,15 +48,13 @@ class ElasticsearchTestCase(TestCase):
         return get_test_client()
 
     @classmethod
-    def setUpClass(cls):
-        super(ElasticsearchTestCase, cls).setUpClass()
+    def setup_class(cls):
         cls.client = cls._get_client()
 
-    def tearDown(self):
-        super(ElasticsearchTestCase, self).tearDown()
+    def teardown_method(self, _):
         # Hidden indices expanded in wildcards in ES 7.7
         expand_wildcards = ["open", "closed"]
-        if self.es_version >= (7, 7):
+        if self.es_version() >= (7, 7):
             expand_wildcards.append("hidden")
 
         self.client.indices.delete(
@@ -65,7 +63,6 @@ class ElasticsearchTestCase(TestCase):
         self.client.indices.delete_template(name="*", ignore=404)
         self.client.indices.delete_index_template(name="*", ignore=404)
 
-    @property
     def es_version(self):
         if not hasattr(self, "_es_version"):
             version_string = self.client.info()["version"]["number"]
