+++ tox.ini
@@ -2,6 +2,8 @@
 envlist = pypy,py27,py34,py35,py36,py37,py38,lint,docs
 [testenv]
 whitelist_externals = git
+deps =
+    -r dev-requirements.txt
 commands =
     python setup.py test
 
@@ -49,7 +51,7 @@ commands =
 
 [testenv:docs]
 deps =
-    sphinx
     sphinx-rtd-theme
+    -r dev-requirements.txt
 commands =
     sphinx-build docs/ docs/_build -b html
